requirements.txt
black==24.10.0
kaggle==1.6.17
numpy==2.1.3
pillow==11.0.0
pytest==8.3.3
ruff==0.6.9
torch==2.5.1
